requirements.txt
yt-dlp
ffmpeg-python
pytesseract
opencv-python
webvtt-py
requests
